Reimplement ZoneHVACComponent::thermalZone and related methods
This is a rework to ZoneHVACComponent::thermalZone ::addToThermalZone
and ::removeFromThermalZone to handle both air based and non air based
equipment. This means equipment that is connected to zone through nodes
as well as equipment that is not. This prevents having to reimplement
these methods in derived classes.

Existing class reimplmentations have been left in place because they are
working, but they might be removed in the future because they are
unnecessary.