Fix partial translation of thermostat dual sp
OS -> E+ translator would happily translate a partial
ThermostatSetpointDualSetpoint even if one or both schedules are
missing. This causes an EnergyPlus simulation failure. This change makes
it so that a dual setpoint thermostat will only be translated if it has
schedules attached.

This change has become more important with the introduction of
ThermalZone grid view, where there is no longer an explicit on/off
button that effectively allows user control over add/remove
ThermostatSetpointDualSetpoint. A different solution to this problem is
to reprogram grid view to remove the thermostat when the last schedule
is removed, but I think this solution will handle more use cases. It
does come at the expense of perhaps a tiny bit of "magic" but I think it
is acceptable, and will certainly keep more simulations running cleanly.

close #1706