SDD: Translate Pump:TotHd directly
Previously the SDD->OS translator only used the TotHd value from the SDD
file if the pump was autosized. Now it will always use that value if
available.

The old logic calculated pump head from power and flow and an assumed
overall efficieincy. That logic is still there but only if the sdd does
not provide a value.

There is now some risk that the head, power, flow, and efficiency values
will over constrain the physical parameters of the pump, but that
checking is now on the CBECC rules engine. E+ will throw an error if the
given values result in a computer efficiency greater than 1.