Add support for design oa spec on terminals
AirTerminalSingleDuctVAVReheat and AirTerminalSingleDuctVAVNoReheat are
the two types of terminals supported by OS that E+ allows to assign a
design oa spec to directly. When doing this it means the terminal will
actively attempt to request flow to meet oa requirements for the
specific zone. This change brings OS up to speed with the E+
capability.

close #1907