Redesign relationship between ControllerWaterCoil and coil
Previously the association between ControllerWaterCoil and the coil was
made by matching the actuator node of the controller to the water inlet
node of the coil. This removes duplication but proves to be fragile.

With this change there is a new field in the controller to explicitly
point to the coil. In the controller the actuator and sensor nodes are
explicitly left blank by coil::addToNode where previously they were set
to appropriate defaults. Now the defaulting is handled in the OS->E+
translator, and only if a user input for these nodes is not already set.
This change should make things more robust when components around the
coil move around.

A relatively complex version translation is required and has been
implemented.

close #1936