return false for table and section now is graceful
Section can be skipped or table within section if a return false is
hit. Warning messages is thrown. For section you know what section is
skipped, for table within section, which table is skipped is unknown.
Ruby error anywhere within the section will skip section with a warning.

Section will be missing from the table of contents in the report if it
is skipped. Could try to alter in the future to issue warning there vs.
hidden in log.