Move version translation for SizingZone to current version
A version translation function delivered on the CBECC branch is longer
valid, merging into current develop. This change is to grab that
translation function for SizingZone and move it to a translate function
pertaining to the current version of OpenStudio.

Editorial Comment: I delivered this to the CBECC branch so that CBECC
models could use new SizingZone features. A version translation was
required and I deliverd that exclusively to the CBECC branch because
develop was in feature freeze. Additionally, delivering these SizingZone
options to develop, and then merging develop into CBECC was not an
option because CBECC was and is currently locked on an E+ version behind
develop and merging develop into CBECC would have the unwanted side
effect of incrementing CBECC's E+ version. Now the downside. There can
be OS models out there generated by CBECC, advertising SizingZone
features that were not supported in mainline OS. This may have tricky
complications if a user attempts to open those models in current (non
cbecc) OpenStudio, because the version translation will effectively be
attempted twice. Bottom line. Keep CBECC branch changes isolated to the
src/sdd directory (aka the CBECC translator) and avoid pusing
exclusive features, especially those that involve version translation,
to CBECC ahead of develop.