Modify ParentObject::remove when curves are involved
ParentObject::remove will remove all children. One problem with this is
that many HVAC objects have curves as children, but now that the curves
are also ResourceObject instances we want them to stick around when you
remove the parent. Here is a change to make that happen. Ignore the
curve children when doing ParentObject::remove.