Add MAKE_POSITION_INDEPENDENT_CODE to geo lib
This basically means that on linux (maybe other compiler configs), cmake
will add the -fPIC flag. This amounts to letting you link a static
library into a shared library, which is the case when we link geolib.a
into the openstudio utilities shared library.

This issue only seems to show up on some linux compiler configurations.
(not mine) and not windows or mac as far as I have found. In any case we
are trusting the cmake abstraction MAKE_POSITION_INDEPENDENT_CODE to do
the right thing.