Remove E+ idd files
Normally when we have an E+ upgrade we archive the old E+ idd and insert
the new one. I think this is unnecessary. OS really only cares about the
proposed idd as far as I know.

It is advantageous to remove these IDD files under the new cycle of
releasing OS in parallel with E+ to avoid having to insert the newest
idd file with the release SHA at the last minute.

The E+ release SHA is still located in the proposed idd and will need to
be maintained.