Make the geolib link private
The link to the static geolib is "public" and therefore transitive. This
means that geolib is linked into utilities, but then when we link
utilities into another project, say resource tests, the geo lib is also
linked into that derived library. Unfortunately the link line in
resource tests (derived project) fails on geo lib because the geo lib
target is not imported to the derived project and therefore unkown.

The solution is to make the link from utlities to geo private, which
breaks this transistive link behavior.