Complete embedded ruby
At this point there is an executable with the full set of ruby libraries
embedded. It is possible to create and save an openstudio model with the
embedded libraries.

exe is about 150mb. it takes about 250 milliseconds to run the exe, init
ruby, init openstudio, create and os model and save. Pretty cool! and
much faster than requiring dynamic openstudio bindings.