Delay require 'openstudio-workflow'
This delays require 'openstudio-workflow' to only the places where it is
needed. This avoids a slight delay if you are just asking for help,
getting version info, using measure manager, etc. It takes about 250 ms
to pull workflow gem off the embedded system and send it through the
ruby parser. Not terrible but it was noticable if you only want to
openstudio -h.

There is still a ~250ms delay due to loading openstudio. This too could
be deferred, but it seems unnecessary for now.