Add link commands for static qt mac
Some system lib link commands were missing. This gets it working,
however there is still a link warning....

It seems that compiling  with -fvisibility=hidden cleans up the warnings
for qt, but then introduces the same type of warning in boost. The
solution is probably to compile all depends with  -fvisibility=hidden.
Meaning recompile boost for mac that way.