Fixes #2240
Basically, there is some real weirdness when multiplying scales.  If the resulting scale cannot be found, the closest one is returned with an adjustment factor, code was ignoring the adjustment factor because there is no way to use it.  I added more scales, this issue can still occur with other units.