Create custom cmake build/install commands for PAT
The rules Download PAT source, extract, npm install, npm install bower,
bower install, npm run release.

An install command is included to put the product into the OS install
tree.

You need node and npm if the build pat option is on. It defaults to off.

Only tested on Mac. I know there will be a couple of path issues that
need to be fixed up on Windows.