Fixing measure manager crash on ETO Measures, cause was measure description including non-ascii character 'temperature set point of 55°F'
This might actually be a larger issue, we have standardized on UTF-8 encoded strings in OpenStudio.  However, it appears that the Ruby C API assumes strings are encoded as ASCII-8BIT. this applies to basically the entire OpenStudio API.  If strings include non ascii characters there may be issues when trying to transcode to another encoding.  We may want to look into a broader fix for this in the future.

https://silverhammermba.github.io/emberb/c/.