Fix problem opening sql in RV from OS App
The issue was that the temp directory where OS puts a copy of the sql
file is actually a symbolic link (this is how the OS sets up the tmp
location). RV has a lot of logic to keep track of open files and
aliases, and some of this logic uses path comparison operations. The
trouble starts when some paths being compared are "canonical" (links
expanded) and others are not expanded. So the solution was just expand
the paths right from the get go. This seemed to only impact files opened
from the cli because it seems the Qt file open dialog already expands
the file paths.

close #2525