List wx as a dependency in the .deb package
The Ubuntu package manager provides wxWidgets as a shared library,
so add it as a dependency to the OpenStudio debian package.
A possible alternative is to build a static library of wx, however
that is a manual task as the package manager does not provide a
static variation of wx.