Fix error finding resources on linux
A fix was recently made to fix errors when calling the cli with a relative path.
That fix was tested and appeared to resolve the issue #2616.

Subsequently another fix was made for calling a sym linked version of cli #2561 (usually Mac)

The sym link fix had a bad interaction, and caused 2615 to reappear.
This commit attempts to reclose #2616 and not break the fix for #2561