CBECC: Update fluid cooler translation
1. EvaporativeFluidCoolerSingleSpeed translate UAFactorSim.
This was done previously, but a recent update changed
```
} else if ( calcMthd.compare("UAFactor", Qt::CaseInsensitive) == 0 ) {
```
to
```
} else if ( calcMthd.compare("UAFactorSim", Qt::CaseInsensitive) == 0 ) {
```
which is incorrect. For calcMthd we still need to look for the enum "UAFactor" for the
correct translation to happen.

2. EvaporativeFluidCoolerTwoSpeed, HtRej:LowSpdPwrRatSim should not have any units,
the incorrect unit translation has been removed.

3. EvaporativeFluidCoolerTwoSpeed, suffered a similiar problem is item 1, for EvaporativeFluidCoolerSingleSpeed
Incorrectly looking for calcMthd "UAFactorSim" instead of "UAFactor". This was introduced by a blanket find/replace.

4. EvaporativeFluidCoolerSingleSpeed, leave  Low Fan Speed U-Factor Times Area Sizing Factor blank

5. FluidCoolerTwoSpeed, add translation for HtRej:LowSpdAirFlowCapSim