Adjust ZoneHVACIdealLoadAirSystem idf translator
If the zones attached to a return plenum contain exactly one
ZoneHVACIdealLoadAirSystem per zone, then in the idf translator
we work around an EnergyPlus bug and connect nodes differently.

Instead of connecting the plenum inlet to a zone exhaust node, we use the zone return
air node. Also, only one of the ZoneHVACIdealLoadAirSystem instances uses the
system inlet node field.

If there are multiple pieces of zone equipment per zone, then all
ZoneHVAC is connected in the conventional way.