Remove large block of if statements in Model_Impl::createObject
Previous REGISTER_CONSTRUCTOR and REGISTER_COPYCONSTRUCTORS macros generated
a very large block of if statements, and for unexplained reasons this lead to
access violations at random points in the createObject process. (on clang)

In place of the if block is a map of IddObjectType to a factory function, which is
statically allocated when the program begins.