Allow SPMs on demand side of plantLoops except demand branches.
Per @kbenne's wishes.

I made a design decision here to leverage polymorphic behavior in order to keep code more centralized at the base class (SetpointManager) level. Instead of overriding the "addToNode" for derived classes (SPMs) that are allowed on PlantLoop to call the SetpointManager_Impl::addToNode and if that fails check if you can connect it to a plantLoop, I created a virtual `isAllowedOnPlantLoop` Implementation method, and whether that returns true or false, the SetpointManager_Impl::addToNode will do the connection to the plantLoop. That way the only override that is needed for SPMs allowed on plantloops is to make `isAllowableOnPlantLoop` return true instead of the default false.
The connection logic stays in one place, so if in the future we want to restrict connection to only the supply side of a plant loop, then there's only one place where the change would be needed instead of multiple files.