Comment out call to _wopen
There seems to be a disconnect between a file opened with `_wopen` and
then later used with SetFilePointer.

However, the win32.c Ruby code was already written to have a fallback to
`CreateFile`, if `_wopen` call failed.

1. This seems to work
2. I have no idea what it *doesn't* fail in Release mode. Maybe file
permissions? Maybe something else?

I had a twitter conversation with one of the MSVC team to try and
determine if my fix seemed reasonable:

https://twitter.com/lefticus/status/1083230026865172480

At the moment, it works. How good of a fix is it? I don't know.