Add link lines for openssh libs on Apple
This is really two things.

1. At the meta project level, compile qt with -openssl-linked,
and be sure to specify the ssl libs found by cmake. project devs should
be sure to have cmake find the brew installed static openssl libraries

2. At the core project level add the brew provided / cmake found openssl libs
to the openstudio cli link path.

All changes are apple specific but may have general application to other platforms.
This is the conservative if it aint broke don't mess with it approach.

ref #104

Thank you @tijcolem for doing the real work on this