CBECC: Use a proper dual setpoint manager when requested
When plant loops were first setup, the option "FixedDualSetpoint" was implemented
using component setpoint operation, at least partly because SetpointManagerScheduledDualSetpoint
was not available in OpenStudio.

Notes on this topic are in the source code:
https://github.com/NREL/OpenStudio/blob/e802fcc20e58aa5778058648f25f75f9ff202bb8/openstudiocore/src/sdd/MapHVAC.cpp#L6196

This change leaves the individal setpoint mangers on each components (which
triggers component setpoint operation scheme in OpenStudio), but changes the setpoint manager
on the supply outlet node from SetpointManagerScheduled to the more expected
SetpointManagerScheduledDualSetpoint.

This is intended to address https://sourceforge.net/p/cbecc-com/tickets/2949/