Fix crash on mac related to os edit
The issue stems from the destructor of OSEdit classes
calling unbind. The unbind method does three things.

1. resets the member variables related to the ModelObject that is bound
2. removes some qt signal connections
3. calls setEnabled(false)

You can see these steps here https://github.com/NREL/OpenStudio/blob/1ae1cec65b25421bd2dabcb7501618a00338645b/openstudiocore/src/shared_gui_components/OSLineEdit.cpp#L152

During destruction of the widget it is item 3 that causes the issue,
because as the widget calls the setEnabled method, the focus event is triggered
and that causes a bunch of refreshing logic that we don't want because the overall
view is in a half destroyed state.

The solution is simply to not call unbind when the OSEdit classes are destroyed.
This is defensbile because

1. member state does not need to be reset because it is soon to be destroyed.
2. qt connections will be removed by QObject destructor https://doc.qt.io/archives/qt-4.8/qobject.html#dtor.QObject
3. setEnable (the source of the issue) does not need to be called because the widget is soon to be
nuked anyway.

close #3434