Add and install openstudioConfigVersion file
This file is installed next to the openstudioConfig file which is used
by consuming cmake projects (such as the app). This is a basic version
file that will allow any consuming projects use any version of
OopenStudio that is equal to or greater than the current version.

For example if the OS SDK (this project) is at version 3.0.0, then
consuming projects may specify

find_project(openstudio 3.0.0) and as long as the SDK is version 3.0.0
or greater the find_project command will succeed.

This is a basic openstudioConfigVersion file. Changes can be made to
make the version compatibility more or less strict.

This is following the cmake documentation here:
https://cmake.org/cmake/help/v3.8/manual/cmake-packages.7.html#creating-packages