Allow embedded gems even with --bundle
Previously, when --bundle is used then even the baked in libraries from openstudio-gems are made unavailable, with this change the baked in openstudio-gems (even native ones like pycall) are still available. Anything defined in --bundle will override the baked in stuff. So for instance it is possible to define a bundle with a custom version of the OpenStudio standards gem and that version of standards will be used instead of the baked in version.

close #3957
close #3933

Also, minor fix to annoying messages shown when NOT using --bundle...
"Ignoring jaro_winkler-1.5.4 because its extensions are not built. Try: gem pristine jaro_winkler --version 1.5.4"
This is not true as jaro_winkler is actually still available from the embedded system. The message has been silenced.