/!\ CHange in ParentObject::remove, so that you remove all children UNLESS they are ResourceObjects and are used
This extends the workaround we had in place for Curves.
Lots of objects list ResourceObjects as children, when they really shouldn't, just because of OSApp InspectorGadget (IG).
This is not an ideal solution (that one would remove the workarounds done for the OSApp only)
Another option would be to not touch this, and have the CoilCoolingDX<...>::remove() call ModelObject_Impl::remove, loop on its resource objects and check if they are still used, if not delete them (I don't think we want to "orphan" Speeds for eg once they become unused).

@kbenne please review this one.