Change parsing of IddFields from back-to-front to front-to-back
 * The IddRegex::lastField is very expensive with any regex processor
 * The code had many expensive string copies to detect end of parsing
 * The fields had to then be reversed after parsing, which is not free

It seems this technique was chosen because it is difficult to detect the
end of a Field, but easy to find the start of one.

However, this function is called many thousands of times during Idd parsing
and the cumulative cost is very high.

To solve the problem we don't try to use a regex to parse the entire field,
instead we hop from beginning to beginning starting at the front.

This makes version translation at least 2x faster due to the savings
in parsing of Idd files

Addresses #637