Pin the actual recipe revisions for all conan dependencies
Pro: we have zero chance of pulling something else. Fully reproducible builds


Cons: whenever we want to update a dependency, we'll have to bump the revision in OpenStudio/ConanInstall.cmake

If we do not pin dependencies, we could simply upload a new recipe revision, and delete the old one from the remote, or pass UPDATE to conan cmake run. I don't like passing UPDATE to conan_cmake_run actually  because it has to query bintray for changes everytime you run cmake (which takes time, doesn't work offline, etc)


Note: the pinning of deps really matter a lot of openstuido_ruby but not so much for the rest of the deps. Because we're unlikely going to bump a dep like boost for no reason without also bumping the version (eg boost/1.74...)