Fix crash in sql with ruby bindings on Windows
Fixes #4109

NOTE: This is currently only tested on Windows!

This crash specifically occurs when using the system Ruby.

sqlite3_initialize was not properly being called in the case where Ruby
is driving the loading of sqlite.

On Windows, DLL's do not share global static values. This is an issue
that has come up in the past in both EnergyPlus and OpenStudio in the
case of singletons.

Fix is to ensure that sqlite3_initialize is called in the DLL

 * Added a global static object that ensures that sqlite3_initialize()
 is called. Hopefully this does not invoke some sort of Static
 Initialization Order Fiasco problem.

Note:

 * I did not get the ruby bindings rebuilding when I changed the
 openstudio_sql.i interface file. Looking into ProjectMacros it seems
 that the KEY_I_FILE is never actually registered as a dependency for
 the SWIG binding being generated. So I added it.