`WorkspaceObject::setString` is currently allowing us to set object names to empty string because it is wrongly assuming that the name string for WorkspaceObject's is always index 0.  For ModelObject's the index is 1 because handle is 0.
Changing this results in some breakage of OutputMeter which uses/abuses the name field to be the E+ meter name.  We may want to change OutputMeter's `Name` field to be `Meter Name`. That seems clear to me but it might break any measure that depends on `meter.name` to give the E+ name. In this commit I kept OutputMeter's `Name` field but marked it as not required (since it is empty) when a meter is first constructed.  This might not be ideal though, we may want to change the field to `Meter Name` to prevent Workspace naming logic from introducing further bugs.

Finally, the Workspace naming logic was protecting the handle field of ModelObjects.  By moving this protection to the name field we may be losing some protection of the handle field.  We may want to add functionality to consider both handle and name fields if they exist?

As with any change to this Workspace level stuff a good review is required.