Add CLI11 for top level argument parsing
If the subcommand `experimental` is passed to the `openstudio` CLI, then
further argument parsing and workflow is diverted through a new C++
based pathway. If the experimental subcommand is not given then the
RubyEngine is loaded and further argument processing is handed of to the
conventional OpenStudio CLI processing and workflow.

ref #4641