Fix ruby+python mixed workflows
One compiler flag - 1 dollar. Knowing which flag to put where,
priceless. (Or at least a couple of weeks of expensive developer time)

This change adds a flag (`-fno-gnu-unique`) to the ruby and python bindings, which is required to make mixed workflows work on GCC. The basic issue is that swig generated bindings contain a bunch of static variables that are overlapping / conflicting on GCC without this flag.

Here is a helpful discussion of a similar issue
https://binutils.sourceware.narkive.com/saHIXe5B/gold-patch-add-no-gnu-unique-option-to-disable-stb-gnu-unique.

close #4719