Fix runtime errors in C++ MeasureManager Server
The MeasureManager server uses the cpprest SDK, which is an asynchronous
library that is implemented on top C++ threads. The challenge is that
API endpoints involving Ruby and Python are not thread safe. (This is
generally true for all conventional implementations of Ruby and Python,
not just OpenStudio's encapsulation of the scripting engines) The
solution implemented here, is to use a task queue so that all of the
"work" done by the API endpoints is done on the common main thread.

ref #4885