Improve OpenStudio's support for Python unitttest
In the Python unittest library, unittest.main() was not working as
expected, because tests defined in the __main__ scope where not being
detected. This meant that `openstudio labs execute_python_script
foo_measure_test.rb` did not work as expected.

The solution here is to register the __main__ module properly within
Python.

* An example Measure test is located at https://github.com/NREL/openstudio-common-measures-gem/commit/0c73cdb0dcde1eac960628e3c4ee853f18ef0f3f
* A Measure test can be invoked with `openstudio labs --python_path=. execute_python_script tests/test_measure.py`

close #4906