Fix runtime error for Python/Ruby workflows
This issue only pertained to Mac and was occuring when a Python Measure was run before a Ruby Measure.
The fix is to set the dlopen flags to RTLD_LOCAL prior to importing OpenStudio in Python.

Test OpenStudioCLI.Labs.Run_PythonRuby demonstrates the original issue and will now pass.

close #4945