Create a new Logger singleton type
Previously, we used the openstudio::Singleton template to create the
singleton named Logger. This implementation suffered a bug described
in #4847, where the singleton was not global across the ScriptEngine DLL
boundary.

This change is a non templated implementation of Logger. The design is
identical to openstudio::Singleton<LoggerSingleton>, but does not suffer
the non-global issue.

ref #4847