Fix error loading gems with classic CLI
This change mirrors https://github.com/NREL/OpenStudio/commit/0fbb223bb9def421962ea1607fcbc5745b88e554,
but applies to the classic CLI. There is code duplication between old
and new implementations of the CLI, which is not resolved by this
commit, however we expect to remove the classic implementation in a
(near?) future OpenStudio release.

This commit also removed embedded_gems.rb, which is an orphaned file
that is not used.