CBECC: Assume 2009 for ScheduleYear
If there are no run period inputs, then we don't have an assumed year.
This change will assume 2009 for the purpose of making ScheduleYear
instances.

This is a light touch and short of a global change to set the
YearDescription::assumedYear, although setting the assumed year is an
alternative path forward.